requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==3.0.1
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
whitenoise==6.9.0
